further deflushing
